ai/ai-engine/requirements.txt
# --- Existing AI Engine Dependencies ---
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
xgboost

# --- NEW: ishaa-ai (Phase 2) Dependencies ---
# Web & DB
motor>=3.4.0
pymongo>=4.7.2

# Caching
redis>=5.0.4
hiredis>=2.3.2  

# AI Providers
google-generativeai>=0.7.0
anthropic>=0.30.0
groq>=0.9.0

# Auth & Crypto
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.8

# Validation & Utils
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
aiofiles>=23.2.1
structlog>=24.1.0